git: pack-redundant
