--- completers/git_completer/cmd/pack_redundant_generated.go
@@ -1,19 +1,0 @@
-package cmd
-
-import (
-	"github.com/carapace-sh/carapace"
-	"github.com/spf13/cobra"
-)
-
-var pack_redundantCmd = &cobra.Command{
-	Use:     "pack-redundant",
-	Short:   "Find redundant pack files",
-	Run:     func(cmd *cobra.Command, args []string) {},
-	GroupID: groups[group_low_level_interrogator].ID,
-}
-
-func init() {
-	carapace.Gen(pack_redundantCmd).Standalone()
-
-	rootCmd.AddCommand(pack_redundantCmd)
-}
